Extract `RejectionDocumentation`
