+++ tools/smoke-tests/rejection-generation/src/test/proto/spine/tools/rejection/rejections.proto
@@ -28,7 +28,7 @@ option (SPI_all) = true;
 option java_package = "io.spine.tools.rejections";
 // Keep all the rejections under the same outer class called `Rejections`.
 
-// The rejection to test creation and validation of a generated rejection.
+// The rejection to test creation and validation of a generated throwable message.
 message CannotUpdateUsername {
 
     // A required name of a user.
